+++ remoulade/state/backends/__init__.py
@@ -9,4 +9,14 @@ except ImportError:  # pragma: no cover
         ImportWarning,
     )
 
-__all__ = ["RedisBackend", "StubBackend"]
+try:
+    from .postgres import PostgresBackend
+except ImportError:  # pragma: no cover
+    import warnings
+
+    warnings.warn(
+        "PostgresBackend is not available.  Run `pip install remoulade[postgres]` " "to add support for that backend.",
+        ImportWarning,
+    )
+
+__all__ = ["RedisBackend", "StubBackend", "PostgresBackend"]