+++ tests/state/test_backend.py
@@ -7,7 +7,9 @@ class TestStateBackend:
 
     def test_change_existence_state(self, state_backend):
         message_id = "14271"
-        state = State(message_id, StateStatusesEnum.Pending, actor_name="do_work", args=[1, 2], kwargs={"name": "work"})
+        state = State(
+            message_id, StateStatusesEnum.Pending, actor_name="do_work", args=[1, 2], kwargs={"status": "work"}
+        )
         state_backend.set_state(state, ttl=100)
         # the state status changes
         state = State(message_id, StateStatusesEnum.Success)
@@ -18,4 +20,4 @@ class TestStateBackend:
         state = state_backend.get_state(message_id)
         assert state.status == StateStatusesEnum.Success
         assert state.args == [1, 2]
-        assert state.kwargs == {"name": "work"}
+        assert state.kwargs == {"status": "work"}