+++ docs/source/changelog.rst
@@ -5,6 +5,21 @@ Changelog
 
 All notable changes to this project will be documented in this file.
 
+`0.39.0`_ -- 2021-09-16
+-----------------------
+Added
+^^^^^
+* state: add new Postgres State Backend
+* compose: add new docker-compose file to easily run all the services necessary to remoulade
+* api: add new /messages/states DELETE route to delete states from the Postgres State Backend
+* api: add new arguments to /messages/states and /groups route to filter results by actor,
+status, id, start_datetime and end_datetime with Postgres State Backend.
+
+Changed
+^^^^^^^
+* api: change /messages/states and /groups routes from GET to POST
+* api: sorting by column is no longer supported with Redis and Stub State Backends
+
 `0.38.1`_ -- 2021-09-16
 -----------------------
 Changed
@@ -723,6 +738,7 @@ Fixed
 .. _#84: https://github.com/wiremind/remoulade/issues/84
 .. _#77: https://github.com/wiremind/remoulade/issues/77
 
+.. _0.39.0: https://github.com/wiremind/remoulade/releases/tag/v0.39.0
 .. _0.38.1: https://github.com/wiremind/remoulade/releases/tag/v0.38.1
 .. _0.38.0: https://github.com/wiremind/remoulade/releases/tag/v0.38.0
 .. _0.37.2: https://github.com/wiremind/remoulade/releases/tag/v0.37.2